# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virta API",
        "description": "API for managing companies and their stations.",
        "contact": {
            "name": "Michael Govern",
            "email": "michaelgovern0826@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/station": {
            "get": {
                "summary": "Get a list of charging stations",
                "description": "Returns a list of all charging stations.",
                "operationId": "2982c0b71cde36d28cc9e1aead019346",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "company_id": {
                                                "type": "integer"
                                            },
                                            "address": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new charging station",
                "description": "Create a new charging station record.",
                "operationId": "75610cdd1a39f87c35d6abd82d96b7db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Station created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "company_id": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/station/{station}": {
            "get": {
                "summary": "Get a specific charging station",
                "description": "Returns details of a specific charging station.",
                "operationId": "bde42f5a3a6cf41badc185c445ab066e",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "ID of the station to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "company_id": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found"
                    }
                }
            },
            "put": {
                "summary": "Update a charging station",
                "description": "Update a charging station record.",
                "operationId": "b956b99f9a497ce369bdff54f664d3a3",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "ID of the station to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Station updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "company_id": {
                                            "type": "integer"
                                        },
                                        "address": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "summary": "Delete a charging station",
                "description": "Delete a charging station record.",
                "operationId": "a0fb02bb5485bd010fd0d1405b54d934",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "ID of the station to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Station deleted successfully"
                    },
                    "404": {
                        "description": "Station not found"
                    }
                }
            }
        },
        "/api/stations-within-radius": {
            "get": {
                "summary": "Get charging stations within a radius",
                "description": "Returns a list of charging stations within a specified radius.",
                "operationId": "fc7140de4e6b3455aa4727ea94f81868",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radius (in kilometers)",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    }
}